+++ src/index.html
@@ -16,7 +16,7 @@
         </div>
         <hr>
         <form id="add-list-form">
-            <input type="text" name="add" placeholder="Add to your list...">
+            <input type="text" name="add" placeholder="Add to your list..." id="addTask" required>
             <button  type="submit" class="btn">
             <i class="fas fa-level-up-alt"></i>
             </button>
