+++ src/index.js
@@ -1,40 +1,10 @@
 import './styles.css';
+import { DisplayTask, addtask } from './StoreList.js';
 
-const tasks = [
-  {
-    descr: 'wash the dishes',
-    completed: true,
-    index: 1,
-  },
-  {
-    descr: 'complete To Do list Project',
-    completed: false,
-    index: 2,
-  },
-];
-
-const taskContainer = document.querySelector('.task-container');
-
-function DisplayTask() {
-  taskContainer.innerHTML = '';
-  const sortedTasks = tasks.sort((x, y) => x.index - y.index);
-  sortedTasks.forEach((task) => {
-    const listItem = document.createElement('li');
-    listItem.innerHTML = `
-            <div class= "container">
-            <div class='task-content'>
-            <input type="checkbox" id="task-${task.index}" ${task.completed ? 'checked' : ''}>
-            <label for="task-${task.index}" class="descr">${task.descr}</label>
-            </div>
-            <div class="icon-container">
-            <i class="fas fa-ellipsis-v"></i>
-            </div>
-            </div>
-            <hr>
-  `;
-    taskContainer.appendChild(listItem);
-  });
-}
+const formBtn = document.querySelector('.btn');
+formBtn.addEventListener('click', (event) => {
+  event.preventDefault();
+  addtask();
+});
 
 DisplayTask();
-// reload Event
